test: move things to more logical homes
