+++ containers/ecr-viewer/tests/unit/app/view-data/components/EcrDocument/EcrDocument.test.tsx
@@ -1,11 +1,24 @@
 import React from "react";
 
+import BundleCareTeam from "@/../../../test-data/fhir/BundleCareTeam.json";
+import BundleWithMiscNotes from "@/../../../test-data/fhir/BundleMiscNotes.json";
+import * as _BundleWithPatient from "@/../../../test-data/fhir/BundlePatient.json";
+import BundleWithPendingResultsOnly from "@/../../../test-data/fhir/BundlePendingResultsOnly.json";
+import BundleWithPlannedMedsOnly from "@/../../../test-data/fhir/BundlePlannedMedsOnly.json";
+import BundleWithScheduledApptsOnly from "@/../../../test-data/fhir/BundleScheduledApptsOnly.json";
 import { render, screen } from "@testing-library/react";
 import { Bundle } from "fhir/r4";
 import { axe } from "jest-axe";
 
+import { DataDisplay } from "@/app/view-data/components/DataDisplay";
 import { EcrDocument } from "@/app/view-data/components/EcrDocument";
 import { getEcrDocumentAccordionItems } from "@/app/view-data/components/EcrDocument/accordion-items";
+import {
+  evaluateClinicalData,
+  returnCareTeamTable,
+} from "@/app/view-data/components/EcrDocument/clinical-data";
+
+const BundleWithPatient = _BundleWithPatient as Bundle;
 
 describe("Snapshot test for eCR Document", () => {
   it("Given no data, info message for empty sections should appear", async () => {
@@ -51,5 +64,84 @@ describe("Snapshot test for eCR Document", () => {
     expect(
       screen.getByText("No eCR metadata was found in this eCR."),
     ).toBeInTheDocument();
+  });
+
+  describe("Evaluate Clinical Info", () => {
+    it("Should return notes", () => {
+      const actual = evaluateClinicalData(
+        BundleWithMiscNotes as unknown as Bundle,
+      );
+      render(actual.clinicalNotes.availableData[0].value as React.JSX.Element);
+      expect(actual.clinicalNotes.availableData[0].title).toEqual(
+        "Miscellaneous Notes",
+      );
+      expect(screen.getByText("Active Problems")).toBeInTheDocument();
+      expect(actual.clinicalNotes.unavailableData).toBeEmpty();
+    });
+    it("Should not include Treatment details if medications is not available", () => {
+      const actual = evaluateClinicalData(
+        BundleWithMiscNotes as unknown as Bundle,
+      );
+      expect(actual.treatmentData.availableData).toBeEmpty();
+    });
+    it("Should return Plan of Treatment when only pending results", () => {
+      const actual = evaluateClinicalData(
+        BundleWithPendingResultsOnly as unknown as Bundle,
+      );
+      expect(actual.treatmentData.availableData[0].title).toEqual(
+        "Plan of Treatment",
+      );
+      const { container } = render(
+        <DataDisplay item={actual.treatmentData.availableData[0]} />,
+      );
+      expect(container).toMatchSnapshot();
+    });
+    it("Should return Plan of Treatment when only scheduled appointments", () => {
+      const actual = evaluateClinicalData(
+        BundleWithScheduledApptsOnly as unknown as Bundle,
+      );
+      expect(actual.treatmentData.availableData[0].title).toEqual(
+        "Plan of Treatment",
+      );
+      const { container } = render(
+        <DataDisplay item={actual.treatmentData.availableData[0]} />,
+      );
+      expect(container).toMatchSnapshot();
+    });
+    it("Should return Plan of Treatment when only ordered meds", () => {
+      const actual = evaluateClinicalData(
+        BundleWithPlannedMedsOnly as unknown as Bundle,
+      );
+      expect(actual.treatmentData.availableData[0].title).toEqual(
+        "Plan of Treatment",
+      );
+      const { container } = render(
+        <DataDisplay item={actual.treatmentData.availableData[0]} />,
+      );
+      expect(container).toMatchSnapshot();
+    });
+  });
+
+  describe("Evaluate Care Team Table", () => {
+    it("should evaluate care team table results", () => {
+      const actual: React.JSX.Element = returnCareTeamTable(
+        BundleCareTeam as unknown as Bundle,
+      ) as React.JSX.Element;
+
+      render(actual);
+
+      expect(screen.getByText("Dr Toob Nix SR")).toBeInTheDocument();
+      expect(screen.getByText("family")).toBeInTheDocument();
+      expect(
+        screen.getByText("Start: 11/16/1884 End: 05/21/1896"),
+      ).toBeInTheDocument();
+    });
+
+    it("the table should not appear when there are no results", () => {
+      const actual = returnCareTeamTable(
+        BundleWithPatient as unknown as Bundle,
+      );
+      expect(actual).toBeUndefined();
+    });
   });
 });
