+++ containers/ecr-viewer/tests/unit/app/view-data/services/evaluateFhirDataServices.test.tsx
@@ -1,12 +1,14 @@
+import BundleEcrMetadata from "@/../../../test-data/fhir/BundleEcrMetadata.json";
+import * as _BundleWithPatient from "@/../../../test-data/fhir/BundlePatient.json";
+import * as _BundleWithDeceasedPatient from "@/../../../test-data/fhir/BundlePatientDeceased.json";
+import BundlePatientMultiple from "@/../../../test-data/fhir/BundlePatientMultiple.json";
+import * as _BundlePatientWithCovid from "@/../../../test-data/fhir/BundlePatientWithCovid.json";
+import BundlePractitionerRole from "@/../../../test-data/fhir/BundlePractitionerRole.json";
+import BundleWithSexualOrientation from "@/../../../test-data/fhir/BundleSexualOrientation.json";
+import BundleWithTravelHistory from "@/../../../test-data/fhir/BundleTravelHistory.json";
 import { render, screen } from "@testing-library/react";
 import { Bundle, BundleEntry, Practitioner } from "fhir/r4";
 
-import BundleEcrMetadata from "../../../../../../../test-data/fhir/BundleEcrMetadata.json";
-import * as _BundleWithPatient from "../../../../../../../test-data/fhir/BundlePatient.json";
-import * as _BundleWithDeceasedPatient from "../../../../../../../test-data/fhir/BundlePatientDeceased.json";
-import BundlePatientMultiple from "../../../../../../../test-data/fhir/BundlePatientMultiple.json";
-import * as _BundlePatientWithCovid from "../../../../../../../test-data/fhir/BundlePatientWithCovid.json";
-import BundlePractitionerRole from "../../../../../../../test-data/fhir/BundlePractitionerRole.json";
 import { formatAge } from "@/app/services/formatService";
 import { evaluateValue } from "@/app/utils/evaluate";
 import mappings from "@/app/utils/evaluate/fhir-paths";
@@ -31,6 +33,7 @@ import {
   evaluateHospitalEncounterData,
   evaluateProviderData,
   evaluatePregnancyData,
+  evaluateSocialData,
 } from "@/app/view-data/services/evaluateFhirDataService";
 
 const BundleWithPatient = _BundleWithPatient as Bundle;
@@ -43,6 +46,32 @@ describe("evaluateFhirDataServices tests", () => {
       const actual = evaluateValue(BundleWithPatient, mappings.patientIds);
 
       expect(actual).toEqual("1234567890");
+    });
+  });
+
+  describe("Evaluate Patient Name", () => {
+    it("should return name", () => {
+      const actual = evaluatePatientName(
+        BundleWithPatient as unknown as Bundle,
+        false,
+      );
+      expect(actual).toEqual("Han Solo");
+    });
+  });
+
+  describe("Extract Patient Address", () => {
+    it("should return empty string if no address is available", () => {
+      const actual = evaluatePatientAddress(undefined as any);
+
+      expect(actual).toBeEmpty();
+    });
+
+    it("should get patient address", () => {
+      const actual = evaluatePatientAddress(
+        BundleWithPatient as unknown as Bundle,
+      );
+
+      expect(actual).toEqual("1 Main St\nCloud City, CA 00000\nUS");
     });
   });
 
@@ -1626,6 +1655,47 @@ Home: 123-456-6909`,
       const actual = evaluatePregnancyData(pregnancyBundle);
       render(actual.availableData[0].value);
       expect(screen.getAllByText("Postpartum Status").length).toEqual(1);
+    });
+  });
+
+  describe("Evaluate Social Data", () => {
+    it("should have no available data when there is no data", () => {
+      const actual = evaluateSocialData(undefined as any);
+
+      expect(actual.availableData).toBeEmpty();
+      expect(actual.unavailableData).not.toBeEmpty();
+    });
+    it("should have travel history when there is a travel history observation present", () => {
+      const actual = evaluateSocialData(
+        BundleWithTravelHistory as unknown as Bundle,
+      );
+
+      render(actual.availableData[0].value);
+      // travel purpose
+      expect(screen.getByText("Active duty military (occupation)"));
+    });
+    it("should have patient sexual orientation when available", () => {
+      const actual = evaluateSocialData(
+        BundleWithSexualOrientation as unknown as Bundle,
+      );
+
+      expect(actual.availableData[0].value).toEqual("Other");
+    });
+    it("should return religion if available", () => {
+      const actual = evaluateSocialData(BundleWithPatient as unknown as Bundle);
+      const ext = actual.availableData.filter(
+        (d) => d.title === "Religious Affiliation",
+      );
+      expect(ext).toHaveLength(1);
+      expect(ext[0].value).toEqual("Baptist");
+    });
+    it("should return marital status if available", () => {
+      const actual = evaluateSocialData(BundleWithPatient as unknown as Bundle);
+      const ext = actual.availableData.filter(
+        (d) => d.title === "Marital Status",
+      );
+      expect(ext).toHaveLength(1);
+      expect(ext[0].value).toEqual("Married");
     });
   });
 });
